Applying All Figma Mockup Content
I have the majority of our mockup applied in here now, if the both of you could refine, check for errors, and fix the buttons to match the button in the navigation bar that would be awesome. I really tried to get this as perfect as I could get it.